+++ lib/forest_liana/engine.rb
@@ -36,6 +36,17 @@ module ForestLiana
       File.basename($0) == 'rake'
     end
 
+    def database_available?
+      database_available = true
+      begin
+        ActiveRecord::Base.connection_pool.with_connection { |connection| connection.active? }
+      rescue => error
+        database_available = false
+        FOREST_LOGGER.error "No Apimap sent to Forest servers, it seems that the database is not accessible:\n#{error}"
+      end
+      database_available
+    end
+
     error = configure_forest_cors unless ENV['FOREST_CORS_DEACTIVATED']
 
     config.after_initialize do |app|
@@ -47,8 +58,10 @@ module ForestLiana
 
         app.eager_load!
 
-        # NOTICE: Do not run the code below on rails g forest_liana:install.
-        Bootstraper.new(app).perform if ForestLiana.env_secret || ForestLiana.secret_key
+        if database_available?
+          # NOTICE: Do not run the code below on rails g forest_liana:install.
+          Bootstraper.new(app).perform if ForestLiana.env_secret || ForestLiana.secret_key
+        end
       end
     end
   end